refactor import proxy logic
